LPS-158949 Refactor useSearchTerm to use also in TeamMembers
